+++ src/Components/CategoriesNew.js
@@ -8,10 +8,6 @@ import { createCategory } from "../Services/Organizations/CategoriesService";
 export default function CategoriesNew(props) {
   const [name, setName] = useState("");
   const { onClose } = props;
-  // const [organizationId, setOrganizationId] = useState("");
-  // const [errors, setErrors] = useState([]);
-  // const [loading, setLoading] = useState(true);
-  // const [isHiddenNew, setIsHiddenNew] = useState(true);
   const {
     currentOrganizationStore,
     currentOrganizationDispatch,
