+++ src/Components/FundingOrgsNew.js
@@ -8,7 +8,7 @@ import { createFundingOrg } from "../Services/Organizations/FundingOrgsService";
 export default function FundingOrgsNew(props) {
   const [name, setName] = useState("");
   const [website, setWebsite] = useState("");
-  const { onClose } = props;
+  const handleCloseFundingOrgsNew = props;
   // const [errors, setErrors] = useState([]);
   const {
     currentOrganizationStore,
@@ -32,7 +32,7 @@ export default function FundingOrgsNew(props) {
           if (fundingOrg) {
             props.updateFundingOrgs(fundingOrg);
             clearForm();
-            onClose();
+            props.handleCloseFundingOrgsNew();
           }
         })
         .catch((error) => {
@@ -43,7 +43,7 @@ export default function FundingOrgsNew(props) {
 
   const handleCancel = (event) => {
     event.preventDefault();
-    onClose();
+    props.handleCloseFundingOrgsNew();
   };
 
   const clearForm = () => {
